Updated Open Source

+++ modules/directory/disciplinereport_delete.php
@@ -17,12 +17,14 @@
     */
 
 	//Required configuration files
-	require(dirname(__FILE__) . '/../../configuration.php');
-	require_once(dirname(__FILE__) . '/../../core/abre_verification.php');
-	require_once('permissions.php');
 
-	$cloudsetting=constant("USE_GOOGLE_CLOUD");
-	if ($cloudsetting=="true") 
+	require_once(dirname(__FILE__) . '/../../core/abre_verification.php');
+	require_once(dirname(__FILE__) . '/../../core/abre_functions.php');
+	require_once('permissions.php');
+	$portal_private_root = getConfigPortalPrivateRoot();
+
+	$cloudsetting = getenv("USE_GOOGLE_CLOUD");
+	if ($cloudsetting=="true")
 		require(dirname(__FILE__). '/../../vendor/autoload.php');
 	use Google\Cloud\Storage\StorageClient;
 
@@ -33,32 +35,34 @@
 
 		//Delete the File
 		include "../../core/abre_dbconnect.php";
-		$sql = "SELECT Filename FROM directory_discipline WHERE id = $id";
+		$sql = "SELECT Filename FROM directory_discipline WHERE id = $id AND siteID = '".$_SESSION['siteID']."'";
 		$result = $db->query($sql);
 		while($row = $result->fetch_assoc()){
 			$filename = htmlspecialchars($row["Filename"], ENT_QUOTES);
 			if($filename != ""){
 				if ($cloudsetting=="true") {
 					$storage = new StorageClient([
-						'projectId' => constant("GC_PROJECT")
-					]);	
-					$bucket = $storage->bucket(constant("GC_BUCKET"));			
+						'projectId' => getenv("GC_PROJECT")
+					]);
+					$bucket = $storage->bucket(getenv("GC_BUCKET"));
 
 					$cloud_dir = "private_html/directory/discipline/" . $filename;
-					$object = $bucket->object($cloud_dir);
-					$object->delete();
+					if($bucket->object($cloud_dir)->exists()){
+						$object = $bucket->object($cloud_dir);
+						$object->delete();
+					}
 				}
 				else {
 					$filename = dirname(__FILE__) . "/../../../$portal_private_root/directory/discipline/" . $filename;
-					unlink($filename);	
+					unlink($filename);
 				}
 			}
 		}
 
 		//Delete the Record
 		include "../../core/abre_dbconnect.php";
-		$stmt = $db->prepare("DELETE from directory_discipline where id = ? LIMIT 1");
-		$stmt->bind_param("i",$id);
+		$stmt = $db->prepare("DELETE FROM directory_discipline WHERE id = ? AND siteID = ? LIMIT 1");
+		$stmt->bind_param("ii",$id, $_SESSION['siteID']);
 		$stmt->execute();
 		$stmt->close();
 		$db->close();
